requirements.txt
streamlit==1.28.0
azure-cognitiveservices-speech==1.34.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
python-dateutil==2.8.2
requests==2.31.0
pydub==0.25.1
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6
streamlit-audiorecorder==0.0.4
